refactor factor
